refactor: add RaceTrackControllerUtils
