+++ src/js/Controllers/RaceTrack/RaceTrackController.js
@@ -1,4 +1,3 @@
-import { createRandomNumber } from "../../utils/utils.js";
 import { getState } from "../../Stores/global/global.js";
 import { subscribe, dispatch, actions, RACE_STATES } from "../../Stores/race/index.js";
 import { raceTrackView } from "../../Views/RaceTrackView/RaceTrackView.js";
@@ -16,8 +15,6 @@ function startRace(raceState) {
   dispatch(actions.DONE);
 }
 
-const MAX_RANDOM_NUMBER = 9;
-const JUDGEMENT_NUMBER = 4;
 const RACE_INTERVAL_TIME = 1000;
 
 function progressRace({ carStates }) {
@@ -25,42 +22,6 @@ function progressRace({ carStates }) {
     const carShouldAdvanceResults = executeNextRace(carStates);
     raceTrackView.continueRace(carShouldAdvanceResults);
   }, RACE_INTERVAL_TIME);
-}
-
-export function executeNextRace(carStates) {
-  const carShouldAdvanceResults = [];
-  const newRaceState = carStates.map((el) => {
-    const shouldAdvance = checkIsCanBeAdvanced(createRandomNumber(MAX_RANDOM_NUMBER));
-    carShouldAdvanceResults.push(shouldAdvance);
-
-    const advanceDistance = getAdvanceDistance(shouldAdvance);
-
-    if (shouldAdvance) {
-      return {
-        ...el,
-        distance: el.distance + advanceDistance,
-      };
-    }
-
-    return el;
-  });
-
-  dispatchNextRaceState(newRaceState);
-
-  return carShouldAdvanceResults;
-}
-
-function dispatchNextRaceState(newRaceState) {
-  dispatch(actions.PROGRESS, newRaceState);
-}
-
-function getAdvanceDistance(shouldAdvance) {
-  if (shouldAdvance) return 1;
-  return 0;
-}
-
-export function checkIsCanBeAdvanced(number) {
-  return number >= JUDGEMENT_NUMBER
 }
 
 subscribe(startRace);
